Extract CsvReader - move DefaultCsvCallback and add module-info
