--- ebean-api/src/main/java/io/ebean/text/csv/package-info.java
@@ -1,4 +1,0 @@
-/**
- * CSV processing objects.
- */
-package io.ebean.text.csv;
