+++ ebean-csv-reader/src/main/java/io/ebean/csv/reader/CsvCallback.java
@@ -1,4 +1,4 @@
-package io.ebean.text.csv;
+package io.ebean.csv.reader;
 
 import io.ebean.Database;
