+++ ebean-csv-reader/src/main/java/io/ebean/csv/reader/DefaultCsvCallback.java
@@ -1,4 +1,4 @@
-package io.ebean.text.csv;
+package io.ebean.csv.reader;
 
 import io.ebean.Database;
 import io.ebean.EbeanVersion;
