+++ ebean-csv-reader/src/main/java/module-info.java
@@ -1,0 +1,6 @@
+module io.ebean.csv.reader {
+
+  requires io.ebean.api;
+
+  exports io.ebean.csv.reader;
+}
