+++ ebean-test/src/test/java/io/ebean/xtest/internal/api/TDSpiServer.java
@@ -13,7 +13,6 @@ import io.ebean.meta.MetaInfoManager;
 import io.ebean.plugin.BeanType;
 import io.ebean.plugin.Property;
 import io.ebean.plugin.SpiServer;
-import io.ebean.text.csv.CsvReader;
 import io.ebean.text.json.JsonContext;
 
 import javax.annotation.Nullable;
